temp-main-branch/requirements.txt
streamlit>=1.38
plotly>=5.24
pandas>=2.2
numpy>=1.26
psycopg2-binary>=2.9
statsmodels>=0.14
python-dotenv>=1.0
